requirements.txt
aiogram==3.*
aiohttp==3.*
python-dotenv==1.0.1